Moved from ACGL to GLOW

+++ code/src/runtime/include/engine/graphics/DrawCall.hpp
@@ -1,17 +1,17 @@
 #pragma once
 #include <engine/graphics/Geometry.hpp>
 #include <engine/graphics/Material.hpp>
-
+#include <glm/gtc/quaternion.hpp>
 
 struct TransformData {
-  glm::dvec3 pos;
+  glm::vec3 pos;
   glm::quat rot;
-  glm::dvec3 scale;
+  glm::vec3 scale;
 };
 
 struct DrawCall {
   Material material;
   Geometry geometry;
-  glm::dmat4 lastRenderTransform;
-  glm::dmat4 thisRenderTransform;
+  glm::mat4 lastRenderTransform;
+  glm::mat4 thisRenderTransform;
 };
